requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
pypdf==5.9.0
python-docx==1.2.0
spacy==3.8.7
sentence-transformers==5.1.0
